+++ src/p_client.cpp
@@ -3333,7 +3333,8 @@ void ClientThink(edict_t *ent, usercmd_t *ucmd)
 
 		if (pm.jump_sound && !(pm.s.pm_flags & PMF_ON_LADDER))
 		{
-			gi.sound(ent, CHAN_VOICE, gi.soundindex("*jump1.wav"), 1, ATTN_NORM, 0);
+			// [Jump] mute jump sounds for viewers with jumpers off
+			Jump_PlayerSound(ent, CHAN_VOICE, gi.soundindex("*jump1.wav"), 1, ATTN_NORM, 0);
 			// Paril: removed to make ambushes more effective and to
 			// not have monsters around corners come to jumps
 			// PlayerNoise(ent, ent->s.origin, PNOISE_SELF);